# Save information about warnings in cache