--- a/scarb/Cargo.toml
+++ b/scarb/Cargo.toml
@@ -16,6 +16,7 @@ repository.workspace = true
 [dependencies]
 anyhow.workspace = true
 async-trait.workspace = true
+bincode.workspace = true
 cairo-lang-compiler.workspace = true
 cairo-lang-defs.workspace = true
 cairo-lang-diagnostics.workspace = true