--- a/scarb/src/compiler/incremental/mod.rs
+++ b/scarb/src/compiler/incremental/mod.rs
@@ -2,4 +2,4 @@ mod compilation;
 mod fingerprint;
 mod source;
 
-pub use compilation::{load_incremental_artifacts, save_incremental_artifacts};
+pub use compilation::{IncrementalContext, load_incremental_artifacts, save_incremental_artifacts};