--- a/scarb/src/compiler/repository.rs
+++ b/scarb/src/compiler/repository.rs
@@ -57,7 +57,7 @@ impl CompilerRepository {
             bail!("unknown compiler for target `{target_kind}`");
         };
         let ctx = load_incremental_artifacts(&unit, db, ws)?;
-        compiler.compile(&unit, ctx.cached_crates(), offloader, db, ws)?;
+        compiler.compile(&unit, &ctx, offloader, db, ws)?;
         save_incremental_artifacts(&unit, db, ctx, ws)?;
         Ok(())
     }