--- a/scarb/src/ops/compile.rs
+++ b/scarb/src/ops/compile.rs
@@ -2,6 +2,7 @@ use crate::compiler::db::{
     ScarbDatabase, build_scarb_root_database, has_plugin, is_starknet_plugin,
 };
 use crate::compiler::helpers::{build_compiler_config, collect_main_crate_ids};
+use crate::compiler::incremental::IncrementalContext;
 use crate::compiler::plugin::proc_macro;
 use crate::compiler::{CairoCompilationUnit, CompilationUnit, CompilationUnitAttributes};
 use crate::core::{
@@ -326,7 +327,13 @@ fn check_unit(unit: CompilationUnit, ws: &Workspace<'_>) -> Result<()> {
                 build_scarb_root_database(&unit, ws, Default::default())?;
             let main_crate_ids = collect_main_crate_ids(&unit, &db);
             check_starknet_dependency(&unit, ws, &db, &package_name);
-            let mut compiler_config = build_compiler_config(&db, &unit, &main_crate_ids, &[], ws);
+            let mut compiler_config = build_compiler_config(
+                &db,
+                &unit,
+                &main_crate_ids,
+                &IncrementalContext::Disabled,
+                ws,
+            );
             let result = compiler_config
                 .diagnostics_reporter
                 .ensure(&db)