This way we can skip calculating diagnostics for main crates.