@maciektr 2025-09-22
Before:
```
Benchmark 1: /Users/maciektr/Projects/scarb/target/release/scarb build -w --test
  Time (mean ± σ):     864.7 ms ±  21.1 ms    [User: 730.4 ms, System: 75.4 ms]
  Range (min … max):   837.8 ms … 887.5 ms    4 runs

Benchmark 1: /Users/maciektr/Projects/scarb/target/release/scarb build -w
  Time (mean ± σ):     455.1 ms ±  11.7 ms    [User: 294.6 ms, System: 34.3 ms]
  Range (min … max):   440.6 ms … 467.7 ms    4 runs

Benchmark 1: /Users/maciektr/Projects/scarb/target/release/scarb build -w --test
  Time (mean ± σ):     39.602 s ±  0.561 s    [User: 76.977 s, System: 6.189 s]
  Range (min … max):   38.901 s … 40.232 s    4 runs

Benchmark 1: /Users/maciektr/Projects/scarb/target/release/scarb build -w
  Time (mean ± σ):     30.995 s ±  0.478 s    [User: 54.296 s, System: 3.481 s]
  Range (min … max):   30.620 s … 31.653 s    4 runs
  ```
  After:
  ```
  Benchmark 1: /Users/maciektr/Projects/scarb/target/release/scarb build -w --test
  Time (mean ± σ):     727.6 ms ±   7.8 ms    [User: 596.9 ms, System: 70.8 ms]
  Range (min … max):   716.1 ms … 733.4 ms    4 runs

Benchmark 1: /Users/maciektr/Projects/scarb/target/release/scarb build -w
  Time (mean ± σ):     402.4 ms ±   5.4 ms    [User: 218.1 ms, System: 31.5 ms]
  Range (min … max):   397.2 ms … 409.4 ms    4 runs

Benchmark 1: /Users/maciektr/Projects/scarb/target/release/scarb build -w --test
  Time (mean ± σ):     36.060 s ±  0.223 s    [User: 57.429 s, System: 4.008 s]
  Range (min … max):   35.881 s … 36.367 s    4 runs

Benchmark 1: /Users/maciektr/Projects/scarb/target/release/scarb build -w
  Time (mean ± σ):     18.767 s ±  0.670 s    [User: 28.766 s, System: 1.859 s]
  Range (min … max):   18.073 s … 19.666 s    4 runs
  ```